+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## v0.1.0 2026-04-01
+
+Added:
+
+- added adversarial network grid, lightcycles and melee
+- added reward and punish to the grid
+Removed:
+
+Improved/Fixed:
+
 # Changelog
 All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
@@ -9,4 +19,3 @@ Added:
 Removed:
 
 Improved/Fixed:
-